python/futu_service/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
futu-api>=9,<10
